// back/chess-ad-server/src/sockets/services/gameplayService/enums.ts
export enum PiecesXpBarrier {
    KNIGHT = 4, 
    BISHOP = 4, 
    ROOK = 6, 
    QUEEN = 8 
  };
  
export enum Pieces { 
  KNIGHT = 'Knight', 
  BISHOP = 'Bishop', 
  ROOK = 'Rook', 
  QUEEN = 'Queen', 
  ZOMBIE = 'Zombie', 
  BROKEN = 'Broken', 
  PAWN = 'Pawn', 
  KING = 'King' 
};

export enum Colors { 
  WHITE = 'white', 
  BLACK = 'black' 
};

export enum ActiveTime { 
  KNIGHT = 5, 
  ZOMBIE = 4,
  BROKEN = 4
};